+++ crates/task-expander/src/token_expander.rs
@@ -589,7 +589,7 @@ impl<'graph> TokenExpander<'graph> {
                 None => Cow::Owned(String::new()),
             },
             "projectChannel" => get_metadata(|md| md.channel.as_deref()),
-            "projectLayer" | "projectType" => Cow::Owned(project.layer.to_string()),
+            "projectLayer" => Cow::Owned(project.layer.to_string()),
             "projectName" => get_metadata(|md| md.name.as_deref()),
             "projectOwner" => get_metadata(|md| md.owner.as_deref()),
             "projectRoot" => Cow::Owned(self.stringify_path(&project.root)?),